evaluation workloads: move data_classes (#286)
* evaluation workloads: move data_classes

* evaluation workloads: pre-commit

* evaluation workloads: rename package
evaluation_workloads: renaming package to llm-evaluation

* evaluation workloads: pre-commit
