+++ docker/llm-evaluation/.dockerignore
