+++ docker/llm-evaluation/run_inference_and_judge_evaluation.py
@@ -3,21 +3,21 @@ import os
 from argparse import Namespace
 from datetime import datetime
 
-from evaluation_metrics import logger
-from evaluation_metrics.argument_parsers import get_judge_inference_parser
-from evaluation_metrics.call_inference_container.call_inference_container import (
+from llm_evaluation import logger
+from llm_evaluation.argument_parsers import get_judge_inference_parser
+from llm_evaluation.call_inference_container.call_inference_container import (
     download_dataset,
     get_llm_client,
     read_prompt_template,
 )
-from evaluation_metrics.call_inference_container.call_inference_container import run as run_call_inference_container
-from evaluation_metrics.call_inference_container.call_inference_container import (
+from llm_evaluation.call_inference_container.call_inference_container import run as run_call_inference_container
+from llm_evaluation.call_inference_container.call_inference_container import (
     save_inference_results,
 )
-from evaluation_metrics.data.data_classes import AggregatedJudgeResults
-from evaluation_metrics.judge.run_judge_evaluation import run_2step_judge_on_inferences
-from evaluation_metrics.metrics.run_metrics_evaluation import read_inference_data
-from evaluation_metrics.metrics.utils import save_results
+from llm_evaluation.data.data_classes import AggregatedJudgeResults
+from llm_evaluation.judge.run_judge_evaluation import run_2step_judge_on_inferences
+from llm_evaluation.metrics.run_metrics_evaluation import read_inference_data
+from llm_evaluation.metrics.utils import save_results
 
 
 async def main(args: Namespace):
