+++ docker/llm-evaluation/src/llm_evaluation/data/data_classes.py
@@ -1,9 +1,28 @@
 import json
 import os
 from dataclasses import dataclass
+from typing import List
 
 from dataclasses_json import dataclass_json
-from evaluation_metrics import logger
+from llm_evaluation import logger
+
+
+@dataclass_json
+@dataclass
+class EvaluationScores:
+    precision_bert: float
+    recall_bert: float
+    f1_bert: float
+    f1_list: List[float]
+    bleu_score: float
+    accuracy: float
+
+
+@dataclass_json
+@dataclass
+class EvaluationResults:
+    prompts: List[str]
+    scores: EvaluationScores
 
 
 @dataclass_json
